Refactors DTOs to Contracts project
Moving the CustomerDto from the Application layer to the Contracts layer, promoting a clearer separation of concerns and adhering to clean architecture principles.

This change ensures that the Application layer does not contain DTOs, which are intended for communication between the Presentation and other layers. The Application layer now depends solely on the Contracts layer for DTO definitions.